+++ src/SFA.DAS.EAS.Portal.Client/DependencyResolution/StructureMap/RecruitApiRegistry.cs
@@ -1,5 +1,5 @@
-using SFA.DAS.EAS.Portal.Client.Http;
-using SFA.DAS.EAS.Portal.Client.Services.DasRecruit;
+using SFA.DAS.EAS.Portal.Client.Services.Recruit;
+using SFA.DAS.EAS.Portal.Client.Services.Recruit.Http;
 using StructureMap;
 
 namespace SFA.DAS.EAS.Portal.Client.DependencyResolution.StructureMap
@@ -9,7 +9,7 @@ namespace SFA.DAS.EAS.Portal.Client.DependencyResolution.StructureMap
         public RecruitApiRegistry()
         {
             For<IRecruitApiHttpClientFactory>().Use<RecruitApiHttpClientFactory>();
-            For<IDasRecruitService>().Use<DasRecruitService>().Singleton();
+            For<IRecruitService>().Use<RecruitService>().Singleton();
         }
     }
 }
