+++ src/SFA.DAS.EAS.Portal.Client/Services/Recruit/Http/IRecruitApiHttpClientFactory.cs
@@ -1,6 +1,6 @@
 using System.Net.Http;
 
-namespace SFA.DAS.EAS.Portal.Client.Http
+namespace SFA.DAS.EAS.Portal.Client.Services.Recruit.Http
 {
     public interface IRecruitApiHttpClientFactory
     {
