+++ src/SFA.DAS.EAS.Portal.Client/Services/Recruit/Http/RecruitApiHttpClientFactory.cs
@@ -4,9 +4,8 @@ using SFA.DAS.EAS.Portal.Client.Configuration;
 using SFA.DAS.Http;
 using SFA.DAS.Http.TokenGenerators;
 
-namespace SFA.DAS.EAS.Portal.Client.Http
+namespace SFA.DAS.EAS.Portal.Client.Services.Recruit.Http
 {
-    //todo: move to sit next to service????
     public class RecruitApiHttpClientFactory : IRecruitApiHttpClientFactory
     {
         private readonly RecruitApiClientConfiguration _recruitApiClientConfig;
