+++ src/SFA.DAS.EAS.Portal/Application/Services/Commitments/Http/ICommitmentsApiHttpClientFactory.cs
@@ -1,6 +1,6 @@
 using System.Net.Http;
 
-namespace SFA.DAS.EAS.Portal.Application.Services.Commitments
+namespace SFA.DAS.EAS.Portal.Application.Services.Commitments.Http
 {
     public interface ICommitmentsApiHttpClientFactory
     {
